pyproject.toml
[tool.poetry]
name = "pocs"
version = "0.1.0"
description = "POCs Ensino"
authors = ["Bruno <bcnbru@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.13"
streamlit = "^1.34.0"
openai = "^1.30.1"
pyarrow = "10.0.1"
streamlit-analytics = "^0.4.1"
openinference-instrumentation-openai = "^0.1.5"
opentelemetry-sdk = "^1.24.0"
opentelemetry-exporter-otlp = "^1.24.0"
st-star-rating = "^0.0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
